There doesn't seem to be ANY way to switch into fullscreen mode using the directx driver.  You need to start there and you can't go back (sigh)

+++ syncterm/conn.c
@@ -1,6 +1,4 @@
-/* Copyright (C), 2007 by Stephen Hurd */
-
-/* $Id: conn.c,v 1.48 2007/11/13 01:37:56 deuce Exp $ */
+/* $Id: conn.c,v 1.46 2007/10/22 20:28:50 deuce Exp $ */
 
 #include <stdlib.h>
 
@@ -29,7 +27,7 @@ char *conn_types[]={"Unknown","RLogin","Telnet","Raw","SSH","Modem"
 ,"Shell"
 #endif
 ,NULL};
-short unsigned int conn_ports[]={0,513,23,0,22,0
+int conn_ports[]={0,513,23,0,22,0
 #ifdef __unix__
 ,65535
 #endif
